+++ src/lib/plugins/dom-commands/video/index.ts
@@ -1,4 +1,4 @@
-import { playVideo, playVideoFactory, videoPlay } from './play-video'
+import { playVideo, playVideoFactory } from './play-video'
 import { clearVideo, clearVideoFactory } from './clear-video';
 import { PluginRegistration } from '../../../base-iv';
 
@@ -9,9 +9,6 @@ export const videoPlugin: PluginRegistration = {
   }, {
     apiName: 'clearVideo',
     apiFn: clearVideo,
-  }, {
-    apiName: 'videoPlay', // Deprecated
-    apiFn: videoPlay,
   }],
   targetFunctionFactories: [
     playVideoFactory,
@@ -23,6 +20,5 @@ declare module '../../../node' {
   interface NodeExtensions {
     clearVideo: typeof clearVideo
     playVideo: typeof playVideo
-    videoPlay: typeof videoPlay
   }
 }